+++ src/routes-nav/RoutesList.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
-import Homepage from "./Homepage";
-import projects from './githubProjects';
+import Homepage from "../Homepage";
+import projects from '../projects/githubProjects';
 
 /** Site-wide routes.
  *
